# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bar Order Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu-items:
    get:
      operationId: getMenuItems
      tags: [menu]
      summary: Get all menu items
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a new menu item (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemBody"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
...
  /order-batches:
    get:
      operationId: getOrderBatches
      tags: [orders]
      summary: Get all order batches
      responses:
        "200":
          description: List of order batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderBatch"
    post:
      operationId: createOrderBatch
      tags: [orders]
      summary: Create a new order batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBatchBody"
      responses:
        "201":
          description: Order batch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBatch"
  /order-batches/direct:
    post:
      operationId: createDirectSale
      tags: [orders]
      summary: Create a sale directly as completed (bartender/admin, no bar display needed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBatchBody"
      responses:
        "201":
          description: Direct sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBatch"
  /order-batches/{id}/edit:
    put:
      operationId: editOrderBatch
      tags: [orders]
      summary: Edit items in an existing order batch (bartender/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditOrderBatchBody"
      responses:
        "200":
          description: Batch updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBatch"
        "404":
          description: Not found
  /order-batches/settle-waiter:
    post:
      operationId: settleWaiterAccount
      tags: [orders]
      summary: Settle all outstanding bills for a specific waiter (bulk mark paid)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettleWaiterBody"
      responses:
        "200":
          description: Account settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettleWaiterResult"
  /order-batches/{id}/return:
    post:
      operationId: returnOrderBatch
      tags: [orders]
      summary: Return a batch to the waitress for correction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnOrderBatchBody"
      responses:
        "200":
          description: Batch returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBatch"
  /order-batches/{id}/resubmit:
    post:
      operationId: resubmitOrderBatch
      tags: [orders]
      summary: Resubmit a returned batch with corrected items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResubmitOrderBatchBody"
      responses:
        "200":
          description: Batch resubmitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBatch"
  /order-batches/{id}/complete:
    post:
      operationId: completeOrderBatch
      tags: [orders]
      summary: Mark an order batch as served (drinks ready)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order batch marked as served
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBatch"
...
  /order-batches/{id}/pay:
    post:
      operationId: payOrderBatch
      tags: [orders]
      summary: Mark an order batch as paid and clear it
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order batch marked as paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBatch"
...
components:
...
  schemas:
...
    MenuItemBody:
      type: object
      required: [name, category, pricePence]
      properties:
        name:
          type: string
        category:
          type: string
        pricePence:
          type: integer
          minimum: 0
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        pricePence:
          type: integer
      required:
        - id
        - name
        - category
        - pricePence
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        pricePence:
          type: integer
      required:
        - id
        - menuItemId
        - menuItemName
        - quantity
        - pricePence
    OrderBatch:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        waitressName:
          type: string
        createdAt:
          type: string
        completedAt:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, completed, paid, returned]
        correctionItemIds:
          type: array
          items:
            type: integer
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - customerName
        - waitressName
        - createdAt
        - status
        - items
    CreateOrderItemBody:
      type: object
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
      required:
        - menuItemId
        - quantity
    CreateOrderBatchBody:
      type: object
      properties:
        customerName:
          type: string
        waitressName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
      required:
        - customerName
        - waitressName
        - items
...
    ResubmitOrderBatchBody:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
    EditOrderBatchBody:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
    SettleWaiterBody:
      type: object
      required: [waitressName]
      properties:
        waitressName:
          type: string
    SettleWaiterResult:
      type: object
      required: [waitressName, count, totalPence]
      properties:
        waitressName:
          type: string
        count:
          type: integer
        totalPence:
          type: integer
    ReturnOrderBatchBody:
      type: object
      required: [correctionItemIds]
      properties:
        correctionItemIds:
          type: array
          items:
            type: integer
...